split out self einnahmen from partner einnahmen
